+++ PomPadDo/Activities/Tasks/TaskRowModifier.swift
@@ -153,6 +153,14 @@ struct TaskRowModifier: ViewModifier {
                 }
             }
             Divider()
+            
+            Button {
+                CalendarManager.addToCalendar(title: task.name, eventStartDate: Date.now, eventEndDate: Date.now)
+            } label: {
+                Label("Add to Calendar", systemImage: "calendar.badge.plus")
+            }
+            
+            Divider()
 
             Menu {
                 ForEach(0...3, id: \.self) { priority in
